# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ApproveDraftsBulkRequest": {
        "description": "Request model for bulk approving drafts.",
        "properties": {
          "draftIds": {
            "items": {
              "type": "string"
            },
            "title": "Draftids",
            "type": "array"
          }
        },
        "required": [
          "draftIds"
        ],
        "title": "ApproveDraftsBulkRequest",
        "type": "object"
      },
      "ApproveDraftsBulkResponse": {
        "description": "Response model for bulk approve operation.",
        "properties": {
          "approvedCount": {
            "title": "Approvedcount",
            "type": "integer"
          },
          "stories": {
            "items": {
              "$ref": "#/components/schemas/BulkApprovedStory"
            },
            "title": "Stories",
            "type": "array"
          }
        },
        "required": [
          "approvedCount",
          "stories"
        ],
        "title": "ApproveDraftsBulkResponse",
        "type": "object"
      },
...
      "BulkApprovedStory": {
        "description": "Lightweight reference to a story created from bulk draft approval.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title"
        ],
        "title": "BulkApprovedStory",
        "type": "object"
      },
...
      "GenerateStoriesRequest": {
        "description": "Request model for generating multiple stories from raw notes.",
        "properties": {
          "projectId": {
            "title": "Projectid",
            "type": "string"
          },
          "rawNotes": {
            "title": "Rawnotes",
            "type": "string"
          }
        },
        "required": [
          "projectId",
          "rawNotes"
        ],
        "title": "GenerateStoriesRequest",
        "type": "object"
      },
      "GenerateStoriesResponse": {
        "description": "Response model for bulk story generation.",
        "properties": {
          "rawNotes": {
            "title": "Rawnotes",
            "type": "string"
          },
          "stories": {
            "items": {
              "$ref": "#/components/schemas/GeneratedStoryResponse"
            },
            "title": "Stories",
            "type": "array"
          }
        },
        "required": [
          "stories",
          "rawNotes"
        ],
        "title": "GenerateStoriesResponse",
        "type": "object"
      },
      "GeneratedStoryResponse": {
        "description": "Response model for a single generated story.",
        "properties": {
          "acceptanceCriteria": {
            "items": {
              "type": "string"
            },
            "title": "Acceptancecriteria",
            "type": "array"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "acceptanceCriteria"
        ],
        "title": "GeneratedStoryResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StoryResponse": {
        "description": "DTO for story response.",
        "properties": {
          "assignedTo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignedto"
          },
          "createdAt": {
            "title": "Createdat",
            "type": "string"
          },
          "createdBy": {
            "title": "Createdby",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "points": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Points"
          },
          "priority": {
            "title": "Priority",
            "type": "string"
          },
          "projectId": {
            "title": "Projectid",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updatedAt": {
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "projectId",
          "createdBy",
          "assignedTo",
          "status",
          "priority",
          "points",
          "createdAt",
          "updatedAt"
        ],
        "title": "StoryResponse",
        "type": "object"
      },
...
      "UpdateStoryDraftRequest": {
        "description": "Request model for updating a story draft.",
        "properties": {
          "acceptanceCriteria": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acceptancecriteria"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "UpdateStoryDraftRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "REST API for Open Projects Hub \u2014 project, client, and story management with JWT authentication and role-based access control. See [docs/api/README.md](https://github.com/alonsovndev/open-projects-hub-api/blob/main/docs/api/README.md) for auth flows, pagination, and code examples.",
    "title": "Open Projects Hub API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/refinement/approve-drafts": {
      "post": {
        "description": "Approve multiple drafts and convert them to stories.\n\nRequires ADMIN role. Bulk operation to approve multiple drafts at once.\n\nArgs:\n    payload: ApproveDraftsBulkRequest with list of draft IDs\n    current_user: Current authenticated admin user\n    use_case: Injected ApproveDraftsBulkUseCase\n\nReturns:\n    ApproveDraftsBulkResponse with approved count and list of created stories\n\nRaises:\n    400: Invalid input or no drafts found\n    500: Internal server error",
        "operationId": "approve_drafts_bulk_v1_refinement_approve_drafts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveDraftsBulkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApproveDraftsBulkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Approve Drafts Bulk",
        "tags": [
          "AI Refinement"
        ]
      }
    },
    "/v1/refinement/drafts/{draft_id}": {
      "patch": {
        "description": "Update a story draft.\n\nRequires ADMIN role. Allows editing the description of an existing draft.\n\nArgs:\n    draft_id: Draft UUID\n    payload: UpdateStoryDraftRequest with fields to update\n    current_user: Current authenticated admin user\n    use_case: Injected UpdateStoryDraftUseCase\n\nReturns:\n    Dict with draft ID on success\n\nRaises:\n    404: Story draft not found\n    500: Internal server error",
        "operationId": "update_draft_v1_refinement_drafts__draft_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStoryDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Update Draft V1 Refinement Drafts  Draft Id  Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Draft",
        "tags": [
          "AI Refinement"
        ]
      }
    },
    "/v1/refinement/drafts/{draft_id}/approve": {
      "post": {
        "description": "Approve a draft and convert it to a story.\n\nRequires ADMIN role. Converts the draft into a full story in the backlog\nand marks the draft as applied.\n\nArgs:\n    draft_id: Draft UUID\n    current_user: Current authenticated admin user\n    use_case: Injected ApproveDraftUseCase\n\nReturns:\n    StoryResponse with created story data\n\nRaises:\n    404: Story draft not found\n    500: Internal server error",
        "operationId": "approve_draft_v1_refinement_drafts__draft_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Approve Draft",
        "tags": [
          "AI Refinement"
        ]
      }
    },
    "/v1/refinement/generate-stories": {
      "post": {
        "description": "Generate story drafts from raw discovery notes using AI.\n\nRequires ADMIN role. Takes raw notes and generates 3-7 well-structured user stories.\n\nArgs:\n    payload: GenerateStoriesRequest with project_id and raw_notes\n    current_user: Current authenticated admin user\n    use_case: Injected GenerateStoriesFromNotesUseCase\n\nReturns:\n    GenerateStoriesResponse with generated stories\n\nRaises:\n    400: Invalid input\n    502: AI service error\n    500: Internal server error",
        "operationId": "generate_stories_v1_refinement_generate_stories_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateStoriesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateStoriesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Generate Stories",
        "tags": [
          "AI Refinement"
        ]
      }
    },
...
  }
}